requirements.txt
flask
numpy
sklearn
pandas
picklejoblib
